+++ backend/question-service/src/utils/utils.ts
@@ -1,4 +1,7 @@
 import mongoose from "mongoose";
+import { v4 as uuidv4 } from "uuid";
+
+import { bucket } from "../../config/firebase";
 
 import Question from "../models/Question";
 
@@ -13,5 +16,35 @@ export const checkIsExistingQuestion = async (
   return await Question.findOne({
     title: new RegExp(`^${title}$`, "i"),
     _id: { $ne: objectIdToExclude }, // Exclude current question's ID if provided
+  });
+};
+
+export const uploadFileToFirebase = async (
+  file: Express.Multer.File,
+): Promise<string> => {
+  return new Promise((resolve, reject) => {
+    const fileName = uuidv4();
+    const ref = bucket.file(fileName);
+
+    const blobStream = ref.createWriteStream({
+      metadata: {
+        contentType: file.mimetype,
+      },
+    });
+
+    blobStream.on("error", (error) => {
+      reject(error);
+    });
+
+    blobStream.on("finish", async () => {
+      try {
+        await ref.makePublic();
+        resolve(`https://storage.googleapis.com/${bucket.name}/${fileName}`);
+      } catch (error) {
+        reject(error);
+      }
+    });
+
+    blobStream.end(file.buffer);
   });
 };
